+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### Fixed
+- Calculation viewport size for Android devices. [Trello 2132](https://trello.com/c/CDfbKUV6)
+
 ## [3.178.0] - 2020-09-10
 ### Updated
 - The `USE_PREDEFINED_DEVICE_INFO` feature now compares device name with all aliases returned from the server. [Trello 301](https://trello.com/c/vGSi2NFz)